Hasta aqui llegamos con cantor

+++ app/src/main/java/com/example/nami/models/detailModels/DeliverConsumerModel.kt
@@ -1,7 +1,7 @@
 package com.example.nami.models.detailModels
 
 data class DeliverConsumerRequest (
-    val idOrder: Long
+    val idOrder: Int
 )
 data class DeliverConsumerResponse (
     val message: String
