+++ app/src/main/java/com/example/nami/models/detailModels/DeliverCourierModel.kt
@@ -1,7 +1,7 @@
 package com.example.nami.models.detailModels
 
 data class DeliverCourierRequest (
-    val idOrder: Long
+    val idOrder: Int
 )
 data class DeliverCourierResponse (
     val message: String
