+++ app/src/main/java/com/example/nami/models/detailModels/PickingModel.kt
@@ -2,15 +2,15 @@ package com.example.nami.models.detailModels
 
 data class PickingOrderRequest (
     val listDataPicker: List<ListDataPicker>,
-    val idOrder: Long,
-    val idUser: Long,
+    val idOrder: Int,
+    val idUser: Int,
     val productosok: Boolean,
     val totalPicker: String,
     val observations: String
 )
 
 data class ListDataPicker (
-    val idDetailOrder: Long,
+    val idDetailOrder: Int,
     val picker: String
 )
