+++ app/src/main/java/com/example/nami/models/detailModels/TakeModel.kt
@@ -5,5 +5,5 @@ data class TakeOrderRequest (
 )
 data class TakeOrderResponse (
     val message: String,
-    val idPickingxOrder: Long
+    val idPickingxOrder: Int
 )
